refactor: server 디렉토리 내 queries, mutations 디렉토리 생성
